0.6.0 remove deprecated stuff
